Fixed radar, smokeScreen, and when difficulty is hard
added a condition for the radar such that if the player has used a smokescreen, the bot should not count the blocks as misses as they might have ships there even if the radar misses. Additionally, the smokeScreen will no longer go to the very first ship, it will go to the first ship that has not been smoked yet. Also, i added a condition to the Fire function such that if it is the bot turn, the bot's public grid will update whether the difficulty is easy or hard.